Move the Sse Notification Service and fix linter in SseNotification resource
