+++ src/main/java/co/edu/itp/svu/web/rest/SseNotificationResource.java
@@ -1,12 +1,11 @@
 package co.edu.itp.svu.web.rest;
 
 import co.edu.itp.svu.security.SecurityUtils;
-import co.edu.itp.svu.service.SseNotificationService;
+import co.edu.itp.svu.service.notification.SseNotificationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
